% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incontinence.R
\docType{data}
\name{incontinence}
\alias{incontinence}
\title{Incontinence example.}
\format{A data frame with 21 observations on the following 4 variables:
\describe{
  \item{y}{Response (1 = continent, 0 = otherwise)}
  \item{\eqn{x_j}}{Lower urinary tract measure, j = 1, 2, 3}
}}
\source{
Heinze, G. (2006). A comparative investigation of methods for logistic
regression with seperated or nearly separated data. \emph{Statistics in Medicine},
25:4216--4226.
}
\usage{
incontinence
}
\description{
A urinary incontinence study.
}
\keyword{datasets}

