% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated.data.R
\docType{data}
\name{simulated.data}
\alias{simulated.data}
\title{We simulate this data for the purpose of illustration of the package}
\format{A data frame with 100 observations with 28 missing responses:}
\usage{
simulated.data
}
\description{
A dataset containing the 100 observations and 4 covariates. The covariates are
generated from standard normal distribution. The missing binary response is
generated using the simulation process as described in the reference.
}
\keyword{datasets}

